tidy up the code
