+++ src/test/scala/uk/gov/hmrc/test/ui/pillar2SubmissionPages/P2SubAgentAffinityKBPage.scala
@@ -14,11 +14,19 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages
+package uk.gov.hmrc.test.ui.pillar2SubmissionPages
 
+import uk.gov.hmrc.test.ui.cucumber.Find.findByCss
 import uk.gov.hmrc.test.ui.cucumber.PageObject
 
-object P2SubBtnFutureAnnualRevenuePage extends PageObject {
-  val url: String                 = s"$p2SubRootUrl"+"below-threshold-notification/future-annual-revenues"
+object P2SubAgentAffinityKBPage extends PageObject {
+  val url: String                 = s"$p2SubRootUrl"
+
+  val header: String  = "h1.govuk-heading-l"
+  val startAgain        = ".govuk-button"
+  val backLink        = ".govuk-back-link"
+
+  def clickStartAgain() = findByCss(startAgain).click()
+
 
 }
