+++ src/test/scala/uk/gov/hmrc/test/ui/pillar2SubmissionPages/P2SubBtnAnnualRevenuePage.scala
@@ -14,13 +14,11 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages
+package uk.gov.hmrc.test.ui.pillar2SubmissionPages
 
 import uk.gov.hmrc.test.ui.cucumber.PageObject
 
-object P2SubmissionSetUpPage extends PageObject {
-
-  val url: String                 = s"$p2SubRootUrl"
-  val header: String              = "h1.govuk-heading-l"
+object P2SubBtnAnnualRevenuePage extends PageObject {
+  val url: String                 = s"$p2SubRootUrl"+"below-threshold-notification/annual-revenues"
 
 }
