+++ src/test/scala/uk/gov/hmrc/test/ui/pillar2SubmissionPages/P2SubBtnFutureAnnualRevenuePage.scala
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages
+package uk.gov.hmrc.test.ui.pillar2SubmissionPages
 
 import uk.gov.hmrc.test.ui.cucumber.PageObject
 
-object P2SubBtnAccountingPage extends PageObject {
-  val url: String                 = s"$p2SubRootUrl"+"below-threshold-notification/accounting-period"
+object P2SubBtnFutureAnnualRevenuePage extends PageObject {
+  val url: String                 = s"$p2SubRootUrl"+"below-threshold-notification/future-annual-revenues"
+
 }
