+++ src/test/scala/uk/gov/hmrc/test/ui/pillar2SubmissionPages/P2SubIndividualAffinityKBPage.scala
@@ -14,11 +14,19 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages
+package uk.gov.hmrc.test.ui.pillar2SubmissionPages
 
+import uk.gov.hmrc.test.ui.cucumber.Find.findByCss
 import uk.gov.hmrc.test.ui.cucumber.PageObject
 
-object P2SubBtnDomesticOrMnePage extends PageObject {
-  val url: String                 = s"$p2SubRootUrl"+"below-threshold-notification/domestic-or-multinational"
+object P2SubIndividualAffinityKBPage extends PageObject {
+  val url: String                 = s"$p2SubRootUrl"
+
+  val header: String  = "h1.govuk-heading-l"
+  val startAgain        = ".govuk-button"
+  val backLink        = ".govuk-back-link"
+
+  def clickStartAgain() = findByCss(startAgain).click()
+
 
 }
