+++ src/test/scala/uk/gov/hmrc/test/ui/pillar2SubmissionPages/P2SubmissionSetUpPage.scala
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2023 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.test.ui.pillar2SubmissionPages
+
+import uk.gov.hmrc.test.ui.cucumber.PageObject
+
+object P2SubmissionSetUpPage extends PageObject {
+
+  val url: String                 = s"$p2SubRootUrl"
+  val header: String              = "h1.govuk-heading-l"
+
+}
