+++ src/test/scala/uk/gov/hmrc/test/ui/pillar2SubmissionPages/P2UkTaxReturnPage.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.test.ui.pages
+package uk.gov.hmrc.test.ui.pillar2SubmissionPages
 
 import uk.gov.hmrc.test.ui.cucumber.PageObject
